S05US22fix

+++ backend/src/models/ActivityManager.js
@@ -1,6 +1,6 @@
 const AbstractManager = require("./AbstractManager");
 
-class UserManager extends AbstractManager {
+class ActivityManager extends AbstractManager {
   constructor() {
     super({ table: "activity" });
   }
@@ -12,21 +12,29 @@ class UserManager extends AbstractManager {
     return result;
   }
 
-  async read(id) {
+  async read(sub) {
     const [result] = await this.database.query(
-      `SELECT job.title, job.type, job.salary, job.description, activity.apply_date AS applyDate, activity.candidate_id AS candidateId FROM ${this.table} LEFT JOIN job ON ${this.table}.job_id = job.id WHERE ${this.table}.candidate_id =?`,
-      [id]
+      `SELECT job.id, job.title, job.type, job.salary, job.description, activity.apply_date AS applyDate, activity.candidate_id AS candidateId FROM ${this.table} LEFT JOIN job ON ${this.table}.job_id = job.id WHERE ${this.table}.candidate_id IN (SELECT id FROM candidate WHERE user_id = ?)`,
+      [sub]
     );
     return result;
   }
 
-  async delete(id) {
+  async create(jobId, candidateId) {
     const [result] = await this.database.query(
-      `DELETE FROM ${this.table} WHERE id=?`,
-      [id]
+      `INSERT INTO ${this.table} (job_id, candidate_id ) VALUES (?, ?)`,
+      [jobId, candidateId]
+    );
+    return result;
+  }
+
+  async delete(id, sub) {
+    const [result] = await this.database.query(
+      `DELETE FROM ${this.table} WHERE job_id = ? AND candidate_id IN (SELECT id FROM candidate WHERE user_id = ?)`,
+      [id, sub]
     );
     return result;
   }
 }
 
-module.exports = UserManager;
+module.exports = ActivityManager;
